requirements.txt
botocore>1.9.20
boto3>1.6.20
Jinja2>2.10
xmltodict>0.11.0
pyyaml
pandas
iso3166
decorator
time

